requirements.txt
api==0.0.7
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
dateparser==1.1.8
defusedxml==0.7.1
Django==3.2.3
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-filters==1.0.0.dev2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Faker==19.3.0
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
lark==1.1.7
MarkupSafe==2.1.3
nose==1.3.7
oauthlib==3.2.2
packaging==23.1
pbr==5.11.1
Pillow==10.0.0
pluggy==0.13.1
psycopg2-binary==2.9.7
py==1.11.0
pycparser==2.21
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
recipe==0.36.5
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
SQLAlchemy==1.3.24
sqlparse==0.4.4
stevedore==5.1.0
structlog==23.1.0
sureberus==0.15.0
tablib==3.5.0
toml==0.10.2
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.4
views==0.3
webcolors==1.11.1
